Fix bug tracking whether `alt` is held down in `Layout` component.
Previously in the `onKeyUp` function in the `Layout` component,
`this.altKeyDown` would **always** end up being set to `false`, no
matter which key was being released. This change only sets
`this.altKeyDown` to `false` if `event.which` indicates that it was the
`alt` key (18) that was released (otherwise, it keeps the state as-is).

This resolves an issue where you could not hold `alt` and press `z`
multiple times to toggle ZenMode.